Trait ink_sandbox::api::contracts_api::ContractAPI

pub trait ContractAPI {
    type T: Config;

    // Required methods
    fn deploy_contract(
        &mut self,
        contract_bytes: Vec<u8>,
        value: <<Self::T as Config>::Currency as Inspect<AccountIdFor<Self::T>>>::Balance,
        data: Vec<u8>,
        salt: Vec<u8>,
        origin: AccountIdFor<Self::T>,
        gas_limit: Weight,
        storage_deposit_limit: Option<<<Self::T as Config>::Currency as Inspect<AccountIdFor<Self::T>>>::Balance>
    ) -> ContractInstantiateResult<AccountIdFor<Self::T>, <<Self::T as Config>::Currency as Inspect<AccountIdFor<Self::T>>>::Balance, EventRecordOf<Self::T>>;
    fn instantiate_contract(
        &mut self,
        code_hash: Vec<u8>,
        value: <<Self::T as Config>::Currency as Inspect<AccountIdFor<Self::T>>>::Balance,
        data: Vec<u8>,
        salt: Vec<u8>,
        origin: AccountIdFor<Self::T>,
        gas_limit: Weight,
        storage_deposit_limit: Option<<<Self::T as Config>::Currency as Inspect<AccountIdFor<Self::T>>>::Balance>
    ) -> ContractInstantiateResult<AccountIdFor<Self::T>, <<Self::T as Config>::Currency as Inspect<AccountIdFor<Self::T>>>::Balance, EventRecordOf<Self::T>>;
    fn upload_contract(
        &mut self,
        contract_bytes: Vec<u8>,
        origin: AccountIdFor<Self::T>,
        storage_deposit_limit: Option<<<Self::T as Config>::Currency as Inspect<AccountIdFor<Self::T>>>::Balance>,
        determinism: Determinism
    ) -> CodeUploadResult<<Self::T as Config>::Hash, <<Self::T as Config>::Currency as Inspect<AccountIdFor<Self::T>>>::Balance>;
    fn call_contract(
        &mut self,
        address: AccountIdFor<Self::T>,
        value: <<Self::T as Config>::Currency as Inspect<AccountIdFor<Self::T>>>::Balance,
        data: Vec<u8>,
        origin: AccountIdFor<Self::T>,
        gas_limit: Weight,
        storage_deposit_limit: Option<<<Self::T as Config>::Currency as Inspect<AccountIdFor<Self::T>>>::Balance>,
        determinism: Determinism
    ) -> ContractExecResultFor<Self::T>;
}

Contract API used to interact with the contracts pallet.

Required Associated Types

type T: Config

The runtime contract config.

Required Methods

fn deploy_contract( &mut self, contract_bytes: Vec<u8>, value: <<Self::T as Config>::Currency as Inspect<AccountIdFor<Self::T>>>::Balance, data: Vec<u8>, salt: Vec<u8>, origin: AccountIdFor<Self::T>, gas_limit: Weight, storage_deposit_limit: Option<<<Self::T as Config>::Currency as Inspect<AccountIdFor<Self::T>>>::Balance> ) -> ContractInstantiateResult<AccountIdFor<Self::T>, <<Self::T as Config>::Currency as Inspect<AccountIdFor<Self::T>>>::Balance, EventRecordOf<Self::T>>

Interface for bare_instantiate contract call with a simultaneous upload.

Arguments

contract_bytes - The contract code.

value - The number of tokens to be transferred to the contract.

data - The input data to be passed to the contract (including constructor name).

salt - The salt to be used for contract address derivation.

origin - The sender of the contract call.

gas_limit - The gas limit for the contract call.

storage_deposit_limit - The storage deposit limit for the contract call.

fn instantiate_contract( &mut self, code_hash: Vec<u8>, value: <<Self::T as Config>::Currency as Inspect<AccountIdFor<Self::T>>>::Balance, data: Vec<u8>, salt: Vec<u8>, origin: AccountIdFor<Self::T>, gas_limit: Weight, storage_deposit_limit: Option<<<Self::T as Config>::Currency as Inspect<AccountIdFor<Self::T>>>::Balance> ) -> ContractInstantiateResult<AccountIdFor<Self::T>, <<Self::T as Config>::Currency as Inspect<AccountIdFor<Self::T>>>::Balance, EventRecordOf<Self::T>>

Interface for bare_instantiate contract call for a previously uploaded contract.

Arguments

code_hash - The code hash of the contract to instantiate.

value - The number of tokens to be transferred to the contract.

data - The input data to be passed to the contract (including constructor name).

salt - The salt to be used for contract address derivation.

origin - The sender of the contract call.

gas_limit - The gas limit for the contract call.

storage_deposit_limit - The storage deposit limit for the contract call.

fn upload_contract( &mut self, contract_bytes: Vec<u8>, origin: AccountIdFor<Self::T>, storage_deposit_limit: Option<<<Self::T as Config>::Currency as Inspect<AccountIdFor<Self::T>>>::Balance>, determinism: Determinism ) -> CodeUploadResult<<Self::T as Config>::Hash, <<Self::T as Config>::Currency as Inspect<AccountIdFor<Self::T>>>::Balance>

Interface for bare_upload_code contract call.

Arguments

contract_bytes - The contract code.

origin - The sender of the contract call.

storage_deposit_limit - The storage deposit limit for the contract call.

fn call_contract( &mut self, address: AccountIdFor<Self::T>, value: <<Self::T as Config>::Currency as Inspect<AccountIdFor<Self::T>>>::Balance, data: Vec<u8>, origin: AccountIdFor<Self::T>, gas_limit: Weight, storage_deposit_limit: Option<<<Self::T as Config>::Currency as Inspect<AccountIdFor<Self::T>>>::Balance>, determinism: Determinism ) -> ContractExecResultFor<Self::T>

Interface for bare_call contract call.

Arguments

address - The address of the contract to be called.

value - The number of tokens to be transferred to the contract.

data - The input data to be passed to the contract (including message name).

origin - The sender of the contract call.

gas_limit - The gas limit for the contract call.

storage_deposit_limit - The storage deposit limit for the contract call.