Enum ink_sandbox::DispatchError

pub enum DispatchError {

    Other(&'static str),
    CannotLookup,
    BadOrigin,
    Module(ModuleError),
    ConsumerRemaining,
    NoProviders,
    TooManyConsumers,
    Token(TokenError),
    Arithmetic(ArithmeticError),
    Transactional(TransactionalError),
    Exhausted,
    Corruption,
    Unavailable,
    RootNotAllowed,

}

Export pallets that are used in crate::create_sandbox Reason why a dispatch call failed.

Variants

Other(&'static str)

Some error occurred.

CannotLookup

Failed to lookup some data.

BadOrigin

A bad origin.

Module(ModuleError)

A custom error in a module.

ConsumerRemaining

At least one consumer is remaining so the account cannot be destroyed.

NoProviders

There are no providers so the account cannot be created.

TooManyConsumers

There are too many consumers so the account cannot be created.

Token(TokenError)

An error to do with tokens.

Arithmetic(ArithmeticError)

An arithmetic error.

Transactional(TransactionalError)

The number of transactional layers has been reached, or we are not in a transactional layer.

Exhausted

Resources exhausted, e.g. attempt to read/write data which is too large to manipulate.

Corruption

The state is corrupt; this is generally not going to fix itself.

Unavailable

Some resource (e.g. a preimage) is unavailable right now. This might fix itself later.

RootNotAllowed

Root origin is not allowed.

Implementations

impl DispatchError

pub fn stripped(self) -> DispatchError

Return the same error but without the attached message.