ink_sandbox

Module macros

Structs

BlockBuilder

A helper struct for initializing and finalizing blocks.

DefaultSandbox

DefaultSandboxRuntime

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

RuntimeOrigin

The runtime origin type representing the origin of a call.

Enums

RuntimeCall

The aggregated runtime call type.

RuntimeEvent

RuntimeHoldReason

A reason for placing a hold on funds.

Type Aliases

Balances

Revive

System

Timestamp