Struct ink_sandbox::macros::RuntimeOrigin

pub struct RuntimeOrigin {
    pub caller: OriginCaller,
    /* private fields */
}

The runtime origin type representing the origin of a call.

Origin is always created with the base filter configured in [$crate::frame_system::Config::BaseCallFilter].

Fields
caller: OriginCaller
Implementations

impl RuntimeOrigin

pub fn none() -> Self

Create with system none origin and [$crate::frame_system::Config::BaseCallFilter].

pub fn root() -> Self

Create with system root origin and [$crate::frame_system::Config::BaseCallFilter].

pub fn signed(by: <DefaultSandboxRuntime as Config>::AccountId) -> Self

Create with system signed origin and [$crate::frame_system::Config::BaseCallFilter].