ink_sandbox

Trait Extension

pub trait Extension: Send + 'static {
    // Required methods
    fn as_mut_any(&mut self) -> &mut (dyn Any + 'static);
    fn type_id(&self) -> TypeId;
}

Export pallets that are used in crate::create_sandbox Marker trait for types that should be registered as Externalities extension.

As extensions are stored as Box<Any>, this trait should give more confidence that the correct type is registered and requested.

Required Methods

fn as_mut_any(&mut self) -> &mut (dyn Any + 'static)

Return the extension as &mut dyn Any.

This is a trick to make the trait type castable into an Any.

fn type_id(&self) -> TypeId

Get the TypeId of this Extension.