ink_sandbox

Trait IntoAccountId

pub trait IntoAccountId<AccountId> {
    // Required method
    fn into_account_id(self) -> AccountId;
}

Trait for types that can be converted into a runtime AccountId.

This allows sandbox APIs to accept various account types without requiring manual conversion.

Required Methods

fn into_account_id(self) -> AccountId