Crate ink_sandbox

Re-exports

pub use macros::BlockBuilder;

pub use macros::DefaultSandbox;

pub use frame_support;

pub use frame_system;

pub use pallet_balances;

pub use pallet_revive;

pub use pallet_timestamp;

pub use paste;

pub use sp_externalities;

Modules

api

macros

Macros

create_sandbox

Macro creating a minimal runtime with the given name. Optionally can take a chain extension type as a second argument.

Structs

AccountId32

Export pallets that are used in crate::create_sandbox An opaque 32-byte cryptographic identifier.

H256

Export pallets that are used in crate::create_sandbox Fixed-size uninterpreted hash type with 32 bytes (256 bits) size.

RuntimeMetadataPrefixed

Metadata prefixed by a u32 for reserved usage

Snapshot

A snapshot of the storage.

Weight

Enums

DispatchError

Export pallets that are used in crate::create_sandbox Reason why a dispatch call failed.

Traits

Extension

Export pallets that are used in crate::create_sandbox Marker trait for types that should be registered as Externalities extension.

Sandbox

Sandbox defines the API of a sandboxed runtime.

Ss58Codec

Export pallets that are used in crate::create_sandbox Key that can be encoded to/from SS58.

Type Aliases

AccountIdFor

Alias for the account ID type.

ContractExecResultFor

Alias for the contract exec result.

ContractInstantiateResultFor

Alias for the contract instantiate result.

ContractResultFor

ContractResultInstantiate

EventRecordOf

Alias for the event record type.

MapAccountResultFor

Alias for the map_acocunt result.

RawStorage

RuntimeCall

Alias for the runtime call type.

RuntimeEventOf

Alias for the runtime event of a sandbox.

RuntimeOf

Alias for the runtime of a sandbox.

StorageRoot

TestExternalities

Export pallets that are used in crate::create_sandbox Type alias for Externalities implementation used in tests.