Trait ink_sandbox::api::system_api::SystemAPI

pub trait SystemAPI {
    type T: Config;

    // Required methods
    fn build_block(&mut self) -> BlockNumberFor<Self::T>;
    fn build_blocks(&mut self, n: u32) -> BlockNumberFor<Self::T>;
    fn block_number(&mut self) -> BlockNumberFor<Self::T>;
    fn events(&mut self) -> Vec<EventRecordOf<Self::T>>;
    fn reset_events(&mut self);
    fn runtime_call<Origin: Into<<RuntimeCall<Self::T> as Dispatchable>::RuntimeOrigin>>(
        &mut self,
        call: RuntimeCall<Self::T>,
        origin: Origin,
    ) -> DispatchResultWithInfo<<RuntimeCall<Self::T> as Dispatchable>::PostInfo>;
}

System API for the sandbox.

Required Associated Types

type T: Config

The runtime system config.

Required Methods

fn build_block(&mut self) -> BlockNumberFor<Self::T>

Build a new empty block and return the new height.

fn build_blocks(&mut self, n: u32) -> BlockNumberFor<Self::T>

Build n empty blocks and return the new height.

Arguments

n - The number of blocks to build.

fn block_number(&mut self) -> BlockNumberFor<Self::T>

Return the current height of the chain.

fn events(&mut self) -> Vec<EventRecordOf<Self::T>>

Return the events of the current block so far.

fn reset_events(&mut self)

Reset the events of the current block.

fn runtime_call<Origin: Into<<RuntimeCall<Self::T> as Dispatchable>::RuntimeOrigin>>( &mut self, call: RuntimeCall<Self::T>, origin: Origin, ) -> DispatchResultWithInfo<<RuntimeCall<Self::T> as Dispatchable>::PostInfo>

Execute a runtime call (dispatchable).

Arguments

call - The runtime call to execute.

origin - The origin of the call.

Object Safety

This trait is not object safe.