ink_sandbox

Type Alias ContractResultInstantiate

pub type ContractResultInstantiate<Runtime> = ContractResult<InstantiateReturnValue, <<Runtime as Config>::Currency as Inspect<AccountIdFor<Runtime>>>::Balance>;

Aliased Type

struct ContractResultInstantiate<Runtime> {
    pub gas_consumed: Weight,
    pub gas_required: Weight,
    pub storage_deposit: StorageDeposit<<<Runtime as Config>::Currency as Inspect<<Runtime as Config>::AccountId>>::Balance>,
    pub result: Result<InstantiateReturnValue, DispatchError>,
}

Fields
gas_consumed: Weight

How much weight was consumed during execution.

gas_required: Weight

How much weight is required as gas limit in order to execute this call.

This value should be used to determine the weight limit for on-chain execution.

Note

This can only be different from [Self::gas_consumed] when weight pre charging is used. Currently, only seal_call_runtime makes use of pre charging. Additionally, any seal_call or seal_instantiate makes use of pre-charging when a non-zero gas_limit argument is supplied.

storage_deposit: StorageDeposit<<<Runtime as Config>::Currency as Inspect<<Runtime as Config>::AccountId>>::Balance>

How much balance was paid by the origin into the contract’s deposit account in order to pay for storage.

The storage deposit is never actually charged from the origin in case of [Self::result] is Err. This is because on error all storage changes are rolled back including the payment of the deposit.

result: Result<InstantiateReturnValue, DispatchError>

The execution result of the wasm code.